subprojects/macos_statusbar/pyproject.toml
[tool.poetry]
name = "hort-statusbar"
version = "0.1.0"
description = "macOS status bar app for openhort — start/stop server, monitor connections, manage plugins"
authors = ["openhort"]
readme = "README.md"
packages = [{include = "hort_statusbar"}]

[tool.poetry.scripts]
hort-statusbar = "hort_statusbar.__main__:main"

[tool.poetry.dependencies]
python = "^3.12"
pyobjc-framework-Cocoa = ">=11.0"
pyobjc-framework-Quartz = ">=11.0"
httpx = ">=0.27"
websockets = ">=13.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
strict = true
python_version = "3.12"
